Use MB_CUR_MAX == 6 by default
Internally, Cygwin already uses __utf8_mbtowc(), even if it still claims
to use the "ASCII" charset.

But the `MB_CUR_MAX` value (which is not actually a constant, but
dependent on the current locale) was still 1, which broke the initial
`globify()` call while parsing the the command-line in `build_argv()`
for non-ASCII arguments.

This fixes https://github.com/git-for-windows/git/issues/2189

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>