nlsfuncs: work around an overzealous GCC warning
GCC 13 (and maybe 12, too), warn about pointers used after `free()`.

In `nlsfuncs.cc`, they are used on purpose, though, in
`rebase_locale_buf()`, to adjust pointers that were invalidated because
of a `realloc()` (not a `free()`, actually).

So let's shush GCC about those instances.

However, we must be careful only to do that with GCC >= 12 because older
versions would complain about an unknown warning...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>